1.因为`Godot`支持了每视口过采样，所以不再需要手动设置字体过采样 2.修改配置文件中显示比例类型为`double`

+++ MVL/Src/Utils/Config/BaseConfig.cs
@@ -12,7 +12,7 @@ public class BaseConfig {
 	public string CurrentModpack { get; set; } = "";
 	public string CurrentAccount { get; set; } = "";
 	public string DisplayLanguage { get; set; } = OS.GetLocale();
-	public float DisplayScale { get; set; } = Tools.GetAutoDisplayScale();
+	public double DisplayScale { get; set; } = Tools.GetAutoDisplayScale();
 	public string ProxyAddress { get; set; } = "";
 	public int DownloadThreads { get; set; } = Environment.ProcessorCount;
 	public string ModpackFolder { get; set; } = Paths.ModpackFolder;
